Added skeleton NodeFactory.
