+++ ArithEval/ArithEval/ArithEval.csproj
@@ -44,6 +44,7 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="Expr.cs" />
+    <Compile Include="INodeFactory.cs" />
     <Compile Include="NodeFactory.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
